#3238 Reworked spell data extraction to use collectors which make things a lot clearer to see and expand upon
