+++ app/Service/CombatLog/DataExtractors/SpellDataCollectors/SpellDataCollectorInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Service\CombatLog\DataExtractors\SpellDataCollectors;
+
+use App\Service\CombatLog\Dtos\DataExtraction\ExtractedDataResult;
+
+interface SpellDataCollectorInterface
+{
+    public function beforeCollect(string $combatLogFilePath): void;
+
+    public function afterCollect(ExtractedDataResult $result, string $combatLogFilePath): void;
+}
